requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.5.0,<4.10
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-decouple>=3.8
slowapi>=0.1.9
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
pytest>=7.0.0
httpx>=0.24.0
